Moves PreAccount into its own file.
