+++ program-runtime/src/lib.rs
@@ -1,10 +1,10 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(min_specialization))]
 
-pub mod instruction_processor;
 pub mod instruction_recorder;
 pub mod invoke_context;
 pub mod log_collector;
 pub mod native_loader;
 pub mod neon_evm_program;
+pub mod pre_account;
 pub mod stable_log;
 pub mod timings;
